[feat] attracion 장소 종류별 검색 기능 추가

+++ bonvoyageW/src/main/java/com/bonvoyage/domain/attraction/entity/AttractionInfoEntity.java
@@ -12,9 +12,9 @@ import javax.persistence.*;
 public class AttractionInfoEntity {
 
     @Id
-    private long contentId;
+    private Long contentId;
 
-    private int contentTypeId;
+    private Long contentTypeId;
     private String title;
     private String addr1;
     private String addr2;
@@ -30,7 +30,7 @@ public class AttractionInfoEntity {
     private String mlevel;
 
     @Builder
-    public AttractionInfoEntity(int contentId, int contentTypeId, String title, String addr1, String addr2, String zipcode, String tel, String firstImage, String firstImage2, int readcount, int sidoCode, int gugunCode, double latitude, double longitude, String mlevel) {
+    public AttractionInfoEntity(Long contentId, Long contentTypeId, String title, String addr1, String addr2, String zipcode, String tel, String firstImage, String firstImage2, int readcount, int sidoCode, int gugunCode, double latitude, double longitude, String mlevel) {
         this.contentId = contentId;
         this.contentTypeId = contentTypeId;
         this.title = title;
